background: moved crono job
The crono job has been moved into the background framework. This way:

- We avoid concurrency problems with registry events and the crono
  synchronization process.
- We can remove a dependency.
- We use only a single process for all background jobs.

This should hopefully also polish some rough edges of this feature.
