+++ Gemfile
@@ -5,7 +5,6 @@ source "https://rubygems.org"
 gem "active_record_union"
 gem "base32"
 gem "bootstrap-sass", "~> 3.3.4"
-gem "crono"
 gem "devise"
 gem "font-awesome-rails", "= 4.7.0.1"
 gem "grape"
