--- config/cronotab.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require "portus/migrate"
-
-env = ENV["CATALOG_CRON"] || 10
-value = Portus::Migrate.from_humanized_time(env, 10)
-Crono.perform(CatalogJob).every value
