+++ db/migrate/20171220095321_drop_crono_jobs_table.rb
@@ -1,0 +1,5 @@
+class DropCronoJobsTable < ActiveRecord::Migration
+  def change
+    drop_table :crono_jobs
+  end
+end
