+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20171213094038) do
+ActiveRecord::Schema.define(version: 20171220095321) do
 
   create_table "activities", force: :cascade do |t|
     t.integer  "trackable_id",   limit: 4
@@ -51,16 +51,6 @@ ActiveRecord::Schema.define(version: 20171213094038) do
 
   add_index "comments", ["repository_id"], name: "index_comments_on_repository_id", using: :btree
   add_index "comments", ["user_id"], name: "index_comments_on_user_id", using: :btree
-
-  create_table "crono_jobs", force: :cascade do |t|
-    t.string   "job_id",            limit: 255, null: false
-    t.datetime "last_performed_at"
-    t.boolean  "healthy"
-    t.datetime "created_at",                    null: false
-    t.datetime "updated_at",                    null: false
-  end
-
-  add_index "crono_jobs", ["job_id"], name: "index_crono_jobs_on_job_id", unique: true, using: :btree
 
   create_table "namespaces", force: :cascade do |t|
     t.string   "name",        limit: 255
