+++ docker-compose.yml
@@ -24,26 +24,6 @@ services:
     volumes:
       - .:/srv/Portus
 
-  crono:
-    image: opensuse/portus:development
-    command: ./bin/crono
-    depends_on:
-      - portus
-      - db
-    environment:
-      - PORTUS_MACHINE_FQDN_VALUE=${MACHINE_FQDN}
-
-      # Theoretically not needed, but cconfig's been buggy on this...
-      - CCONFIG_PREFIX=PORTUS
-
-      - PORTUS_DB_HOST=db
-      - PORTUS_DB_PASSWORD=portus
-      - PORTUS_DB_POOL=5
-    volumes:
-      - .:/srv/Portus
-    links:
-      - db
-
   background:
     image: opensuse/portus:development
     entrypoint: bundle exec rails runner /srv/Portus/bin/background.rb
