+++ packaging/suse/portus.spec.in
@@ -128,9 +128,7 @@ rm -rf %{buildroot}/%{portusdir}/tmp
 mkdir %{buildroot}/%{portusdir}/tmp
 
 install -d %{buildroot}/%{_sbindir}
-cp packaging/suse/bin/portus_crono %{buildroot}/%{_sbindir}/
 cp packaging/suse/bin/portusctl %{buildroot}/%{_sbindir}/
-%__install -D -m 444 packaging/suse/conf/portus_crono.service %{buildroot}%{_unitdir}/portus_crono.service
 
 # Man pages
 install -d %{buildroot}%{_mandir}/man1
@@ -139,13 +137,11 @@ install -p -m 644 packaging/suse/portusctl/man/man1/*.1 %{buildroot}%{_mandir}/m
 %fdupes %{buildroot}/%{portusdir}
 
 %pre
-%service_add_pre portus_crono.service
 
 %post
 if [ -d /srv/Portus/tmp ];then
   chown -R wwwrun:www /srv/Portus/tmp
 fi
-%service_add_post portus_crono.service
 
 if [ \! -e "/srv/Portus/config/config-local.yml" ]; then
   # First installation of Portus
@@ -153,10 +149,8 @@ if [ \! -e "/srv/Portus/config/config-local.yml" ]; then
 fi
 
 %preun
-%service_del_preun portus_crono.service
 
 %postun
-%service_del_postun portus_crono.service
 
 %files
 %defattr(-,root,root)
@@ -184,9 +178,7 @@ fi
 %{portusdir}/log/
 %{portusdir}/tmp/
 %{portusdir}/db/
-%{_sbindir}/portus_crono
 %{_sbindir}/portusctl
-%{_unitdir}/portus_crono.service
 %{_mandir}/man1/portusctl-*.1.gz
 %{_mandir}/man1/portusctl.1.gz
