+++ spec/lib/portus/background/sync_spec.rb
@@ -1,0 +1,197 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+require "portus/background/sync"
+
+# Just opens up protected methods so they can be used in the test suite.
+class SyncMock < ::Portus::Background::Sync
+  def update_registry!(catalog)
+    super
+  end
+end
+
+describe ::Portus::Background::Sync do
+  describe "#sleep_value" do
+    it "returns always 20" do
+      expect(subject.sleep_value).to eq 20
+    end
+  end
+
+  describe "#work?" do
+    it "always return true" do
+      expect(subject.work?).to be_truthy
+    end
+  end
+
+  describe "#execute!" do
+    describe "Empty database" do
+      it "updates the registry" do
+        allow_any_instance_of(::Portus::RegistryClient).to receive(:manifest).and_return(["", ""])
+
+        registry  = create(:registry)
+        namespace = create(:namespace, registry: registry)
+
+        sync = SyncMock.new
+        sync.update_registry!([{ "name" => "busybox", "tags" => ["latest", "0.1"] },
+                               { "name" => "alpine", "tags" => ["latest"] },
+                               { "name" => "#{namespace.name}/alpine", "tags" => ["latest"] }])
+
+        # Global repos.
+        ns = Namespace.where(global: true)
+        repos = Repository.where(namespace: ns)
+        expect(repos.map(&:name).sort).to match_array(%w[alpine busybox])
+        tags = Repository.find_by(name: "busybox").tags
+        expect(tags.map(&:name)).to match_array(["0.1", "latest"])
+        tags = Repository.find_by(name: "alpine", namespace: ns).tags
+        expect(tags.map(&:name)).to match_array(["latest"])
+
+        # Local repos.
+        repos = Repository.where(namespace: namespace)
+        expect(repos.map(&:name).sort).to match_array(["alpine"])
+        tags = repos.first.tags
+        expect(tags.map(&:name)).to match_array(["latest"])
+      end
+
+      it "does nothing if there is no registry" do
+        sync = SyncMock.new
+        expect { sync.execute! }.not_to raise_error
+      end
+
+      it "raises an exception when there has been a problem in /v2/_catalog" do
+        VCR.turn_on!
+
+        create(:registry, "hostname" => "registry.test.lan")
+
+        VCR.use_cassette("registry/get_missing_catalog_endpoint", record: :none) do
+          sync = SyncMock.new
+          expect(Rails.logger).to receive(:warn).with(/page not found/)
+          sync.execute!
+        end
+      end
+
+      it "executes the task as expected" do
+        VCR.turn_on!
+
+        allow_any_instance_of(::Portus::RegistryClient).to receive(:manifest).and_return(["", ""])
+        registry = create(:registry, "hostname" => "registry.test.lan")
+
+        VCR.use_cassette("registry/get_registry_catalog", record: :none) do
+          sync = SyncMock.new
+          sync.execute!
+        end
+
+        repos = Repository.all
+        expect(repos.count).to eq 1
+        repo = repos[0]
+        expect(repo.name).to eq "busybox"
+        expect(repo.namespace.id).to eq registry.namespaces.first.id
+        tags = repo.tags
+        expect(tags.map(&:name)).to match_array(["latest"])
+      end
+
+      it "handles registries even if there some namespaces missing" do
+        VCR.turn_on!
+
+        allow_any_instance_of(::Portus::RegistryClient).to receive(:manifest).and_return(["", ""])
+        registry = create(:registry, "hostname" => "registry.test.lan")
+
+        VCR.use_cassette("registry/get_registry_catalog_namespace_missing", record: :none) do
+          sync = SyncMock.new
+          sync.execute!
+        end
+
+        repos = Repository.all
+        expect(repos.count).to eq 1
+        repo = repos[0]
+        expect(repo.name).to eq "busybox"
+        expect(repo.namespace.id).to eq registry.namespaces.first.id
+        tags = repo.tags
+        expect(tags.map(&:name)).to match_array(["latest"])
+      end
+    end
+
+    describe "Database already filled with repos" do
+      let!(:registry)    { create(:registry) }
+      let!(:owner)       { create(:user) }
+      let!(:namespace)   { create(:namespace, registry: registry) }
+      let!(:repo1)       { create(:repository, name: "repo1", namespace: namespace) }
+      let!(:repo2)       { create(:repository, name: "repo2", namespace: namespace) }
+      let!(:tag1)        { create(:tag, name: "tag1", repository: repo1) }
+      let!(:tag2)        { create(:tag, name: "tag2", repository: repo2) }
+      let!(:tag3)        { create(:tag, name: "tag3", repository: repo2) }
+
+      it "updates the registry" do
+        allow_any_instance_of(::Portus::RegistryClient).to receive(:manifest).and_return(["", ""])
+
+        sync = SyncMock.new
+        sync.update_registry!([{ "name" => "busybox", "tags" => ["latest", "0.1"] },
+                               { "name" => "#{namespace.name}/repo1",  "tags" => ["latest"] },
+                               { "name" => "#{namespace.name}/repo2",  "tags" => %w[latest tag2] },
+                               { "name" => "#{namespace.name}/alpine", "tags" => ["latest"] }])
+
+        # Global repos
+        ns = Namespace.where(global: true)
+        repos = Repository.where(namespace: ns)
+        expect(repos.map(&:name).sort).to match_array(["busybox"])
+        tags = repos.first.tags
+        expect(tags.map(&:name).sort).to match_array(["0.1", "latest"])
+
+        # User namespaces.
+        repos = Repository.where(namespace: namespace)
+        expect(repos.map(&:name).sort).to match_array(%w[alpine repo1 repo2])
+        tags = Repository.find_by(name: "alpine").tags
+        expect(tags.map(&:name)).to match_array(["latest"])
+        tags = Repository.find_by(name: "repo1").tags
+        expect(tags.map(&:name)).to match_array(["latest"])
+        tags = Repository.find_by(name: "repo2").tags
+        expect(tags.map(&:name)).to match_array(%w[latest tag2])
+      end
+    end
+
+    describe "uploading repository whose tags is nil" do
+      it "skip this repository" do
+        sync = SyncMock.new
+        sync.update_registry!([{ "name" => "busybox", "tags" => nil }])
+
+        # Global repos
+        ns = Namespace.where(global: true)
+        repos = Repository.where(namespace: ns)
+        expect(repos.length).to eq 0
+      end
+    end
+
+    describe "Activities are updated accordingly" do
+      let!(:registry)   { create(:registry) }
+      let!(:owner)      { create(:user) }
+      let!(:repo)       { create(:repository, name: "repo", namespace: registry.global_namespace) }
+      let!(:tag)        { create(:tag, name: "latest", author: owner, repository: repo) }
+
+      it "removes activities from dangling tags" do
+        allow_any_instance_of(::Portus::RegistryClient).to receive(:manifest).and_return(["", ""])
+
+        repo.create_activity(:push, owner: owner, recipient: tag)
+
+        activities = PublicActivity::Activity.order(:updated_at)
+        expect(activities.count).to eq 1
+        expect(activities.first.parameters[:tag_name]).to be_nil
+
+        sync = SyncMock.new
+        sync.update_registry!([{ "name" => "repo", "tags" => ["0.1"] }])
+
+        # Three activities: the original one, one more push for 0.1 and then the
+        # delete for latest.
+        activities = PublicActivity::Activity.order(:updated_at)
+        expect(activities.count).to eq 3
+        expect(activities.first.parameters[:tag_name]).to eq "latest"
+        expect(activities[1].recipient.name).to eq "0.1"
+        expect(activities.last.key).to eq "repository.delete"
+      end
+    end
+  end
+
+  describe "#to_s" do
+    it "works" do
+      expect(subject.to_s).to eq "Registry synchronization"
+    end
+  end
+end
